second

+++ api/views.py
@@ -1,13 +1,13 @@
 from django.contrib.auth import get_user_model
 from django.http import JsonResponse
-from django.shortcuts import get_object_or_404
-from rest_framework import viewsets, filters, mixins
+from recipes.models import Favorite, Follow, Ingredient
+from rest_framework import filters, mixins, viewsets
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
+from shopping_list.mixins import ShopList
 
-from recipes.models import Ingredient, Follow, Favorite, Recipe
-
-
-from .serializers import (IngredientSerializer)
+from .serializers import (FavoriteSerializer, IngredientSerializer,
+                          SubscribeSerializer)
 
 User = get_user_model()
 
@@ -22,35 +22,57 @@ class IngredientViewSet(mixins.ListModelMixin, viewsets.GenericViewSet):
     search_fields = ['^name', ]
 
 
-class SubscribeView(APIView):
-    def post(self, request):
-        author_id = int(self.request.data.get('id'))
-        author = get_object_or_404(User, id=author_id)
-        user = request.user
-        if author_id and author != user:
-            Follow.objects.get_or_create(user=user, author=author)
-            return SUCCESS_RESPONSE
-        else:
-            return BAD_RESPONSE
+class SubscribeViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
+    queryset = Follow.objects.all()
+    permission_classes = (IsAuthenticated,)
+    serializer_class = SubscribeSerializer
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+
+class FavoriteViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
+    queryset = Favorite.objects.all()
+    permission_classes = (IsAuthenticated,)
+    serializer_class = FavoriteSerializer
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+
+class SubscribeDeleteViewSet(APIView):
+    permission_classes = (IsAuthenticated,)
 
     def delete(self, request, author_id):
         follow = Follow.objects.filter(user=request.user, author=author_id)
-        follow.delete()
-        return SUCCESS_RESPONSE
-
-
-class FavoriteView(APIView):
-    def post(self, request):
-        recipe_id = int(self.request.data.get('id'))
-        recipe = get_object_or_404(Recipe, id=recipe_id)
-        user = request.user
-        if recipe_id:
-            Favorite.objects.get_or_create(user=user, recipe=recipe)
+        if follow:
+            follow.delete()
             return SUCCESS_RESPONSE
-        else:
-            return BAD_RESPONSE
+        return BAD_RESPONSE
+
+
+class FavoriteDeleteViewSet(APIView):
+    permission_classes = (IsAuthenticated,)
 
     def delete(self, request, recipe_id):
-        Favorite.objects.filter(
-            user=request.user, recipe=recipe_id).delete()
+        favorite = Favorite.objects.filter(
+            user=request.user, recipe=recipe_id)
+        if favorite:
+            favorite.delete()
+            return SUCCESS_RESPONSE
+        return BAD_RESPONSE
+
+
+class ShopListViewSet(APIView):
+    def post(self, request):
+        shop_list = ShopList(request)
+        recipe_id = self.request.data.get('id')
+        if recipe_id is not None:
+            shop_list.add(int(recipe_id))
+            return SUCCESS_RESPONSE
+        return BAD_RESPONSE
+
+    def delete(self, request, recipe_id):
+        shop_list = ShopList(request)
+        shop_list.remove(int(recipe_id))
         return SUCCESS_RESPONSE
